Fagområdetabell i egen komponent
